feat: prototype of home menu
