--- src/components/ServerIcon.css
@@ -1,19 +1,0 @@
-.guild-icon {
-    background-color: var(--bg-1);
-    cursor: pointer;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 24px;
-
-    width: 48px;
-    height: 48px;
-    border-radius: 50%;
-
-    transition: border-radius 90ms ease-in-out;
-}
-
-button.guild-icon {
-    border: none;
-    color: var(--green);
-}
