+++ src/components/home/HomeContent.svelte
@@ -1,0 +1,24 @@
+<script lang="ts">
+    import IconParkOutlineSleepOne from "~icons/icon-park-outline/sleep-one";
+</script>
+
+<div class="lonely">
+    <div class="lonely-icon">
+        <IconParkOutlineSleepOne />
+    </div>
+    <div>*crickets*</div>
+</div>
+
+<style>
+    .lonely {
+        text-align: center;
+        font-weight: 600;
+        font-size: 32px;
+        color: var(--lonely-message-fg);
+        user-select: none;
+    }
+    .lonely-icon {
+        line-height: 0;
+        font-size: 128px;
+    }
+</style>
